html/semantics/document-metadata/the-meta-element/encoding.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Character encoding</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/document-metadata.html#charset">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=characters>foo&#x263A;bar&#x1F4A9;</div>
<div id="log"></div>
<script>
  test(function(){
    assert_equals(document.getElementById("characters").textContent, "foo☺bar💩");
  });
</script>
